Cleanup and grammar updates. Finished porting the parser module.
- Some adjustments for XCode 8. Fixed also a few warnings.
- Removed old grammars and updated build scripts for macOS and Win to create the ANTLR parser files too.
- Finished porting the parser module to the ANTLR4 parser. Unit tests succeed.
- Further rework of the MySQL grammar:
  * consistent naming
  * a few more changes to stay closer to sql_yacc.yy.

+++ library/base/file_utilities.cpp
@@ -280,7 +280,7 @@ namespace base {
     {
       char pid[32];
       // couldn't lock file, check if we own it ourselves
-      int c = read(fd, pid, sizeof(pid)-1);
+      ssize_t c = read(fd, pid, sizeof(pid)-1);
       close(fd);
       if (c < 0)
         return LockedOther;
